+++ src/core/custom-button/custom-button.styled.tsx
@@ -3,13 +3,29 @@ import styled from 'styled-components/native'
 import { COLOR } from '../colors/colors.constants'
 import { TEXT_SIZE } from '../text/text.constants'
 
-export const CustomButtonContainer = styled.TouchableOpacity`
+interface CustomButtonContainerProps {
+  invisible?: boolean;
+  color?: COLOR;
+}
+
+const getCustomButtonContainerOpacity = R.pipe(
+  R.propOr(false, 'invisible'),
+  R.not,
+  Number,
+)
+
+const getCustomButtonContainerBackgroundColor = R.propOr(COLOR.DODGER_BLUE, 'color')
+
+export const CustomButtonContainer = styled.TouchableOpacity.attrs<CustomButtonContainerProps>(({ invisible }) => ({
+  activeOpacity: invisible ? 0 : undefined,
+}))`
   justify-content: center;
   align-items: center;
   align-self: center;
   padding: 10px 15px;
   border-radius: 30px;
-  background-color: ${COLOR.DODGER_BLUE};
+  background-color: ${getCustomButtonContainerBackgroundColor};
+  opacity: ${getCustomButtonContainerOpacity};
 `
 
 const getCustomButtonTitleFontSize = R.propOr(TEXT_SIZE.M, 'titleSize')