+++ src/dashboard/dashboard.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { Host } from 'react-native-portalize'
-import { Button, View } from 'react-native'
+import { View } from 'react-native'
 import { useSelectLetter } from './hooks/use-select-letter.hook'
 import { PossibleWordsModal } from './components/possible-words-modal/possible-words-modal'
 import { LettersSlider } from '../core/letters-slider/letters-slider'
@@ -8,13 +8,17 @@ import { SelectedLetters } from './components/selected-letters/selected-letters'
 import { LettersGrid } from './components/letters-grid/letters-grid'
 import { CustomButton } from '../core/custom-button/custom-button'
 import { useSearchPossibleWords } from './hooks/use-search-possible-words.hook'
-import { DashboardSafeArea, SearchButtonIcon } from './dashboard.styled'
+import { COLOR } from '../core/colors/colors.constants'
+import {
+  ClearLettersButtonIcon, DashboardButtonsContainer, DashboardSafeArea, SearchButtonIcon,
+} from './dashboard.styled'
 
 export const Dashboard = () => {
   const modalizeRef = React.useRef<any>(null)
 
   const {
-    letters, selectedLetters, handleSelectLetter, handleDeselectLetter, soapCharactersIndexes,
+    letters, selectedLetters,
+    handleSelectLetter, handleDeselectLetter, soapCharactersIndexes, handleClearSelectedLetters,
   } = useSelectLetter()
 
   const {
@@ -36,10 +40,27 @@ export const Dashboard = () => {
         <View>
           <LettersGrid {...{ letters, handleSelectLetter }} />
           <LettersSlider onChange={onLengthChange} />
-          <CustomButton onPress={onSearch}>
-            <SearchButtonIcon />
-          </CustomButton>
 
+          <DashboardButtonsContainer>
+            <CustomButton onPress={() => null} invisible>
+              <SearchButtonIcon />
+            </CustomButton>
+
+            <CustomButton
+              invisible={selectedLetters.length < 2}
+              onPress={onSearch}
+            >
+              <SearchButtonIcon />
+            </CustomButton>
+
+            <CustomButton
+              color={COLOR.FIRE_BRICK}
+              invisible={!selectedLetters.length}
+              onPress={handleClearSelectedLetters}
+            >
+              <ClearLettersButtonIcon />
+            </CustomButton>
+          </DashboardButtonsContainer>
         </View>
 
         <PossibleWordsModal